simplify testing of create_runpath
